@@ -2,8 +2,8 @@ namespace CowLibrary
 {
     public abstract class SceneObject
     {
-        public int Id { get; set; }
+        public int Id { get; init; }
 
-        public Transform Transform { get; set; } = new Transform();
+        public Transform Transform { get; init; } = new();
     }
 }
