@@ -13,7 +13,7 @@ namespace CowLibrary
 
         public override float Evaluate(float cosThetaI)
         {
-            cosThetaI = Mathf.Clamp(cosThetaI, -1, 1);
+            cosThetaI = Math.Clamp(cosThetaI, -1, 1);
             var eta = etaT / etaI;
             var etaK = k / etaI;
 
